Move tests to appropriate folder
